refactor: refactor reader routines to separate file
