+++ src/sgs_tools/io/sgs.py
@@ -1,9 +1,11 @@
 from pathlib import Path
-from typing import Any
+from typing import Any, Dict
 
 import xarray as xr
 
-from sgs_tools.io.um import restrict_ds
+from sgs_tools.io.read_util import restrict_ds, standardize_varnames
+
+degenerate_naming_convention: Dict[str, str] = {}
 
 
 def data_ingest_SGS(
@@ -12,7 +14,7 @@ def data_ingest_SGS(
     chunks: Any = "auto",
 ):
     """read and pre-process local-convention (sgs_tools) NetCDF data using sgs_tools naming convention.
-    Any unknown fields will retain their original names.
+    Will not rename any fields, assume they are in local convention.
 
     :param fname_pattern: NetCDF diagnostic file to read. can be a glob pattern. (should belong to the same simulation)
     :param requested_fields: list of fields to retain in ds, if falsy will retain all.
@@ -25,6 +27,9 @@ def data_ingest_SGS(
     )
 
     ds = xr.open_mfdataset(fname, chunks=chunks, parallel=True, engine="h5netcdf")
+    # rename to sgs_tools naming convention
+    ds = standardize_varnames(ds, degenerate_naming_convention)
+
     if requested_fields:
         ds, _ = restrict_ds(ds, requested_fields)
     return ds
